apps/backend/worker/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0
unstructured[pdf]==0.12.0
unstructured-inference==0.7.0
pdf2image==1.17.0
Pillow==10.2.0
sentence-transformers==2.3.1
numpy==1.26.3
tenacity==8.2.3
